# your_project/docs/swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
...
  UsersCount:
    description: UsersCount represents the count of users.
    properties:
      count:
        description: Count is the number of users.
        type: integer
    type: object
  UsersDetails:
    description: UsersDetails represents detailed information about a specific user.
    properties:
      country:
        description: Country is the country of residence of the user.
        type: string
      createdAt:
        description: CreatedAt is the timestamp indicating when the user entry was
          created.
        type: string
      email:
        description: Email is the email address of the user.
        type: string
      firstname:
        description: Firstname is the first name of the user.
        type: string
      id:
        description: ID is the unique identifier for the user.
        type: string
      lastname:
        description: Lastname is the last name of the user.
        type: string
      status:
        description: Status is a boolean indicating the current status of the user.
        type: boolean
    type: object
  UsersIn:
    description: UsersIn represents the input structure for creating a new user.
    properties:
      companyID:
        description: CompanyID is the unique identifier for the company associated
          with the user. It is required.
        type: string
      email:
        description: Email is the email address of the user. It is required, should
          be a valid email, and maximum length is 255 characters.
        maxLength: 255
        type: string
      firstName:
        description: Firstname is the first name of the user. It is required and should
          be between 3 and 30 characters.
        maxLength: 30
        minLength: 3
        type: string
      lastName:
        description: Lastname is the last name of the user. It is required and should
          be between 3 and 35 characters.
        maxLength: 35
        minLength: 3
        type: string
      password:
        description: Password is the user's password. It is required, and its length
          should be between 10 and 255 characters.
        maxLength: 255
        minLength: 10
        type: string
    required:
    - companyID
    - email
    - firstName
    - lastName
    - password
    type: object
  UsersList:
    description: UsersList represents a simplified version of the user for listing
      purposes.
    properties:
      id:
        description: ID is the unique identifier for the user.
        type: string
      name:
        description: Name is the full name of the user.
        type: string
    type: object
  UsersPagination:
    description: UsersPagination represents the paginated list of users.
    properties:
      items:
        description: Items is a slice containing individual user details.
        items:
          $ref: '#/definitions/UsersTable'
        type: array
      limit:
        description: Limit is the maximum number of items per page in the pagination.
        type: integer
      page:
        description: Page is the current page number in the pagination.
        type: integer
      totalCount:
        description: TotalCount is the total number of users in the entire list.
        type: integer
    type: object
  UsersTable:
    description: UsersTable represents a single user entry in a table.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the user entry was
          created.
        type: string
      email:
        description: Email is the email address of the user.
        type: string
      firstname:
        description: Firstname is the first name of the user.
        type: string
      id:
        description: ID is the unique identifier for the user.
        type: string
      lastname:
        description: Lastname is the last name of the user.
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /users/{companyID}:
    get:
      description: Get all users.
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UsersPagination'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: User query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UsersIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create user
      tags:
      - Users
  /users/{companyID}/{ID}:
    delete:
      description: Delete one user.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: User ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      description: Get one user.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: User ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UsersDetails'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: User ID
        in: path
        name: ID
        required: true
        type: string
      - description: User query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UsersIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
  /users/{companyID}/count:
    get:
      description: Get number of all users.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UsersCount'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get number of  users
      tags:
      - Users
  /users/{companyID}/list:
    get:
      description: Get list of all users.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UsersList'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get list of  users
      tags:
      - Users
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"